Adding Forex Model
