+++ app/config/cryptoCompare.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   KEY: "3ed405fdb692a8718752990780fb5831f176121289496731fadffe3d28bf670f",
-  TTL: 120
+  TTL: 60
 };
