+++ app/config/finnHub.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   KEY: "c4bflciad3i848j2b7ig",
-  TTL: 120
+  TTL: 60
 };
